fix int type conversion warning

+++ src/ops/linear/cpu/linear_cpu.cpp
@@ -14,7 +14,7 @@ void _linear(T *out, const T *in, const T *weight, const T* bias,
         k,        // lda
         k,        // ldb
         n,        // ldc
-        [bias] (float acc, size_t im, size_t in) -> T{
+        [bias] (float acc, int64_t im, int64_t in) -> T{
             (void)im;
             if (bias != nullptr) {
                 acc += llaisys::utils::cast<float>(bias[in]);
